Refactor rake tasks
