+++ .solargraph.yml
@@ -1,6 +1,7 @@
 ---
 include:
   - "**/*.rb"
+  - "**/*.rake"
   - Rakefile
 exclude:
   - "spec/**/*"
