+++ lib/task_helpers/util.rb
@@ -6,7 +6,7 @@ require "semantic"
 
 module TaskUtil
   def root
-    @root ||= File.expand_path("..", __dir__)
+    @root ||= File.expand_path("../..", __dir__)
   end
 
   def adaptors
