+++ lib/tasks/release.rake
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "util"
+require_relative "../task_helpers/util"
 
 class ReleaseTasks
   include Rake::DSL
